NEED CONFIRM: relocate io.aims test helper functions
