+++ tests/io/aims/sets/test_md_generator.py
@@ -3,13 +3,13 @@
 from __future__ import annotations
 
 import pytest
+from aims.conftest import Si, compare_files
 
 from pymatgen.io.aims.sets.core import MDSetGenerator
 from pymatgen.util.testing import TEST_FILES_DIR
-from pymatgen.util.testing.aims import Si, compare_files
 
 SPECIES_DIR = TEST_FILES_DIR / "io/aims/species_directory"
-REF_PATH = (TEST_FILES_DIR / "io/aims/aims_input_generator_ref").resolve()
+REF_PATH = TEST_FILES_DIR / "io/aims/aims_input_generator_ref"
 
 
 def test_si_md(tmp_path):
