+++ tests/io/aims/sets/test_static_restart_from_relax_generator.py
@@ -2,9 +2,10 @@
 
 from __future__ import annotations
 
+from aims.conftest import O2, Si, comp_system
+
 from pymatgen.io.aims.sets.core import StaticSetGenerator
 from pymatgen.util.testing import TEST_FILES_DIR
-from pymatgen.util.testing.aims import O2, Si, comp_system
 
 SPECIES_DIR = TEST_FILES_DIR / "io/aims/species_directory"
 REF_PATH = (TEST_FILES_DIR / "io/aims/aims_input_generator_ref").resolve()
